Apply ZRound convert in single page view model.
